reorganized modals into folder
